+++ ouds_core/lib/components/utilities/markdown_span_builder.dart
@@ -145,6 +145,7 @@ class MarkdownSpanBuilder {
             style: baseStyle.copyWith(
               fontWeight: FontWeight.bold,
               decoration: TextDecoration.underline,
+              decorationColor: baseStyle.color,
             ),
           ),
         );